added dynamic stock watching

+++ stock-montior/src/main/java/org/jamescarr/stocks/StockMonitor.java
@@ -9,7 +9,7 @@ public class StockMonitor {
 	private static final String[] TICKERS = {"IBM", "ORCL", "GOOG"};
 	private final StockService stockService;
 	private final StockNotifier notifier;
-	private final ConcurrentMap<String, Float> previousValue = new ConcurrentHashMap<String, Float>();
+	private final ConcurrentMap<String, Double> previousValue = new ConcurrentHashMap<String, Double>();
 	
 	public StockMonitor(StockService stockService, StockNotifier notifier) {
 		this.stockService = stockService;
@@ -20,16 +20,16 @@ public class StockMonitor {
 	@Scheduled(fixedRate=2000)
 	public void poll(){
 		for(String ticker : TICKERS){
-			Float price = stockService.getQuote(ticker);
+			double price = stockService.getQuote(ticker);
 			notifier.stockPrice(new StockPrice(ticker, price, getPreviousValue(ticker)));
 			previousValue.put(ticker, price);
 		}
 	}
 
 
-	private Float getPreviousValue(String ticker) {
-		final Float previous = previousValue.get(ticker);
-		return previous != null? previous : 0.00f;
+	private Double getPreviousValue(String ticker) {
+		final Double previous = previousValue.get(ticker);
+		return previous != null? previous : 0.00;
 	}
 
 }
