+++ stock-montior/src/main/java/org/jamescarr/stocks/StockPrice.java
@@ -1,31 +1,32 @@
 package org.jamescarr.stocks;
 
+
 public class StockPrice {
-	private final float price;
-	private final float previous;
+	private final double price;
+	private final double previous;
 	private final String ticker;
 
-	public StockPrice(String ticker, float price, float previous) {
+	public StockPrice(String ticker, double price, double previous) {
 		this.ticker = ticker;
 		this.price = price;
 		this.previous = previous;
 	}
 
-	public float getPrevious() {
+	public double getPrevious() {
 		return previous;
 	}
 
-	public float getPrice() {
-		return price;
+	public double getPrice() {
+		return price ;
 	}
 
 	public String getTicker() {
 		return ticker;
 	}
 
-	public Float getChange() {
-		Float difference = price - previous;
-		return new Float(Math.round(difference*100.0) / 100.0);
+	public double getChange() {
+		double difference = price - previous;
+		return new Double(Math.round(difference*100.0) / 100.0);
 	}
